Corrected the uSync page URL casing
If you click on the project URL when viewing the NuGet package in Visual Studio you get a 404 error.  That's because it's using https://jumoo.co.uk/uSync which doesn't work, the URL needs to be https://jumoo.co.uk/usync (lower case u).  I have therefore updated the URL with the lower case u to fix that.